Moved sync to pkg/sync2
