+++ pkg/sync2/syncmap.go
@@ -4,33 +4,35 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-package typ
+package sync2
 
 import (
 	"sync"
 	"sync/atomic"
 	"unsafe"
+
+	"gopkg.in/typ.v3"
 )
 
-// SyncMap is like a Go map[K]V but is safe for concurrent use
+// Map is like a Go map[K]V but is safe for concurrent use
 // by multiple goroutines without additional locking or coordination.
 // Loads, stores, and deletes run in amortized constant time.
 //
-// The SyncMap type is specialized. Most code should use a plain Go map instead,
+// The Map type is specialized. Most code should use a plain Go map instead,
 // with separate locking or coordination, for better type safety and to make it
 // easier to maintain other invariants along with the map content.
 //
-// The SyncMap type is optimized for two common use cases: (1) when the entry for a given
+// The Map type is optimized for two common use cases: (1) when the entry for a given
 // key is only ever written once but read many times, as in caches that only grow,
 // or (2) when multiple goroutines read, write, and overwrite entries for disjoint
-// sets of keys. In these two cases, use of a SyncMap may significantly reduce lock
+// sets of keys. In these two cases, use of a Map may significantly reduce lock
 // contention compared to a Go map paired with a separate Mutex or RWMutex.
 //
-// The zero SyncMap is empty and ready for use. A SyncMap must not be copied after first use.
+// The zero Map is empty and ready for use. A Map must not be copied after first use.
 //
 // This map is a fork of the official Go implementation, with the change of
 // making it generic.
-type SyncMap[K comparable, V any] struct {
+type Map[K comparable, V any] struct {
 	mu sync.Mutex
 
 	// read contains the portion of the map's contents that are safe for
@@ -106,7 +108,7 @@ func newEntry[T any](i T) *entry[T] {
 // Load returns the value stored in the map for a key, or nil if no
 // value is present.
 // The ok result indicates whether value was found in the map.
-func (m *SyncMap[K, V]) Load(key K) (value V, ok bool) {
+func (m *Map[K, V]) Load(key K) (value V, ok bool) {
 	read, _ := m.read.Load().(readOnly[K, V])
 	e, ok := read.m[key]
 	if !ok && read.amended {
@@ -126,7 +128,7 @@ func (m *SyncMap[K, V]) Load(key K) (value V, ok bool) {
 		m.mu.Unlock()
 	}
 	if !ok {
-		return Zero[V](), false
+		return typ.Zero[V](), false
 	}
 	return e.load()
 }
@@ -134,13 +136,13 @@ func (m *SyncMap[K, V]) Load(key K) (value V, ok bool) {
 func (m *entry[T]) load() (value T, ok bool) {
 	p := atomic.LoadPointer(&m.p)
 	if p == nil || p == expunged {
-		return Zero[T](), false
+		return typ.Zero[T](), false
 	}
 	return *(*T)(p), true
 }
 
 // Store sets the value for a key.
-func (m *SyncMap[K, V]) Store(key K, value V) {
+func (m *Map[K, V]) Store(key K, value V) {
 	read, _ := m.read.Load().(readOnly[K, V])
 	if e, ok := read.m[key]; ok && e.tryStore(&value) {
 		return
@@ -203,7 +205,7 @@ func (m *entry[T]) storeLocked(i *T) {
 // LoadOrStore returns the existing value for the key if present.
 // Otherwise, it stores and returns the given value.
 // The loaded result is true if the value was loaded, false if stored.
-func (m *SyncMap[K, V]) LoadOrStore(key K, value V) (actual V, loaded bool) {
+func (m *Map[K, V]) LoadOrStore(key K, value V) (actual V, loaded bool) {
 	// Avoid locking if it's a clean hit.
 	read, _ := m.read.Load().(readOnly[K, V])
 	if e, ok := read.m[key]; ok {
@@ -246,7 +248,7 @@ func (m *SyncMap[K, V]) LoadOrStore(key K, value V) (actual V, loaded bool) {
 func (m *entry[T]) tryLoadOrStore(i T) (actual T, loaded, ok bool) {
 	p := atomic.LoadPointer(&m.p)
 	if p == expunged {
-		return Zero[T](), false, false
+		return typ.Zero[T](), false, false
 	}
 	if p != nil {
 		return *(*T)(p), true, true
@@ -262,7 +264,7 @@ func (m *entry[T]) tryLoadOrStore(i T) (actual T, loaded, ok bool) {
 		}
 		p = atomic.LoadPointer(&m.p)
 		if p == expunged {
-			return Zero[T](), false, false
+			return typ.Zero[T](), false, false
 		}
 		if p != nil {
 			return *(*T)(p), true, true
@@ -272,7 +274,7 @@ func (m *entry[T]) tryLoadOrStore(i T) (actual T, loaded, ok bool) {
 
 // LoadAndDelete deletes the value for a key, returning the previous value if any.
 // The loaded result reports whether the key was present.
-func (m *SyncMap[K, V]) LoadAndDelete(key K) (value V, loaded bool) {
+func (m *Map[K, V]) LoadAndDelete(key K) (value V, loaded bool) {
 	read, _ := m.read.Load().(readOnly[K, V])
 	e, ok := read.m[key]
 	if !ok && read.amended {
@@ -292,11 +294,11 @@ func (m *SyncMap[K, V]) LoadAndDelete(key K) (value V, loaded bool) {
 	if ok {
 		return e.delete()
 	}
-	return Zero[V](), false
+	return typ.Zero[V](), false
 }
 
 // Delete deletes the value for a key.
-func (m *SyncMap[K, V]) Delete(key K) {
+func (m *Map[K, V]) Delete(key K) {
 	m.LoadAndDelete(key)
 }
 
@@ -304,7 +306,7 @@ func (m *entry[T]) delete() (value T, ok bool) {
 	for {
 		p := atomic.LoadPointer(&m.p)
 		if p == nil || p == expunged {
-			return Zero[T](), false
+			return typ.Zero[T](), false
 		}
 		if atomic.CompareAndSwapPointer(&m.p, p, nil) {
 			return *(*T)(p), true
@@ -322,7 +324,7 @@ func (m *entry[T]) delete() (value T, ok bool) {
 //
 // Range may be O(N) with the number of elements in the map even if f returns
 // false after a constant number of calls.
-func (m *SyncMap[K, V]) Range(f func(key K, value V) bool) {
+func (m *Map[K, V]) Range(f func(key K, value V) bool) {
 	// We need to be able to iterate over all of the keys that were already
 	// present at the start of the call to Range.
 	// If read.amended is false, then read.m satisfies that property without
@@ -355,7 +357,7 @@ func (m *SyncMap[K, V]) Range(f func(key K, value V) bool) {
 	}
 }
 
-func (m *SyncMap[K, V]) missLocked() {
+func (m *Map[K, V]) missLocked() {
 	m.misses++
 	if m.misses < len(m.dirty) {
 		return
@@ -365,7 +367,7 @@ func (m *SyncMap[K, V]) missLocked() {
 	m.misses = 0
 }
 
-func (m *SyncMap[K, V]) dirtyLocked() {
+func (m *Map[K, V]) dirtyLocked() {
 	if m.dirty != nil {
 		return
 	}
